Makefile.glyphs: Remove some unnecessary override possibilities
Some apps are setting `GLYPH_SRC_DIR` to `glyphs` which create a
circular dependency and doesn't bring much now that generated file sare
put in a dedicated folder by default instead of in the src folder.
`GLYPH_PATH` is no longer supported, apps need to put `glyphs` in
`glyphs/` folder. This is not used in any apps, so safe to remove.