lib_standard_app: main.c: Change TRY/CATCH behavior to exit the app
Previous behavior was to catch the Exception and somehow restart the
app. This was somehow hiding the error as the user would not see the
app crash. As Exception should not be raised in nominal situation or
should be caught before, we changed the behavior to directly exit the
app.

This also have the benefit of avoiding issue where the whole app BSS
should be reset to have a sane behavior when the app was restarted.

(cherry picked from commit 3c83546929746cdf0dfb01973c19f4ea3dd05854)