Fat: Extend compression to smaller images
Work on glyphs generation script.

Smaller images were excluded from compression (because compression is
incompatible with geometry transformation, and smaller images are the
only ones being transformed). This commit adds the possibility to exclude an image from compression by suffixing it with
"_nocomp". This allow to compress images that were not compressed
before.