lib_ux/include/ux.h: Fix LNX and LNSP version of UX_WAIT_DISPLAYED()
On devices where the SE drives the screen, there is no need for seph
exchanges to control the screen.
The previous version was assuming UX_DISPLAY_NEXT_ELEMENT() was changing
the state of the MCU <-> SE exchanges, which is false, this was leading
to a deadlock on Speculos.

(cherry picked from commit 1b75b439278c93d4cd1428dc6ed5b9d5e3215f16)