lib_cxng: Silent -Wunused-result warnings
This warnings were raised since the recent addition of
WARN_UNUSED_RESULT in the CX lib function prototypes.

For the cx_hash_xxx() calls, an error can be returned only in situation
where an invalid parameter is passed, which will be catched earlier.

Yet this has been fixed more properly here:
https://github.com/LedgerHQ/ledger-secure-sdk/commit/ba40e20d89b2f93ecafcf10a24234734cce9bd39#diff-860e74f69399a4dbbfb6[…]1add4a8e6f2606747eb865fc5e
But this shouldn't be done on this SDK as this would change the real
implementation of the CX functions that are included in the OS and
therefore won't be align with the SDK source if we were to change them.

That is why we choosed to silent the warnings.